+++ src/protocols/rdp/keymaps/failsafe.keymap
