+++ src/protocols/rdp/keymaps/fr_ch_qwertz.keymap
