+++ src/protocols/rdp/keymaps/fr_fr_azerty.keymap
