+++ src/protocols/rdp/keymaps/ja_jp_qwerty.keymap
